% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addRecaptures}
\alias{addRecaptures}
\title{Add recapture locations to an existing plot}
\usage{
addRecaptures(
  Signal,
  shape = 21,
  size = 1.5,
  colour = "white",
  fill = "dodgerblue"
)
}
\arguments{
\item{Signal}{The signal of the transmitter of interest}

\item{shape}{The shape of the points}

\item{size}{The size of the points}

\item{colour}{The colour of the points}

\item{fill}{The fill of the points}
}
\value{
A ggplot with the recapture locations
}
\description{
Add recapture locations to an existing plot
}
