% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addStations}
\alias{addStations}
\title{Add receiver stations to an existing plot}
\usage{
addStations(input, shape = 21, size = 1.5, colour = "white", fill = "black")
}
\arguments{
\item{input}{The output of \code{\link{runRSP}} or \code{\link{dynBBMM}}}

\item{shape}{The shape of the points}

\item{size}{The size of the points}

\item{colour}{The colour of the points}

\item{fill}{The fill of the points}
}
\value{
A ggplot with stations
}
\description{
Add receiver stations to an existing plot
}
\examples{
\donttest{
# Import river shapefile
water <- actel::shapeToRaster(shape = paste0(system.file(package = "RSP"), "/River_latlon.shp"), 
size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Run dynamic Brownian Bridge Movement Model (dBBMM)
dbbmm.data <- dynBBMM(input = rsp.data, base.raster = water, UTM = 56)

# Plot example dBBMM with acoustic stations
plotContours(dbbmm.data, tag = "A69-9001-1111", track = 1) + addStations(rsp.data)
}

}
