% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_helpers.R
\name{breakByTimeframe}
\alias{breakByTimeframe}
\title{Split detections every time "timeframe" hours pass.}
\usage{
breakByTimeframe(input, timerange, timeframe)
}
\arguments{
\item{input}{a list of detections per group}

\item{timerange}{the time-range over which detections are to be used}

\item{timeframe}{the interval of hours in each time slot (in hours)}
}
\value{
the detections split by group and time slot
}
\description{
Split detections every time "timeframe" hours pass.
}
\keyword{internal}
