% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRSP.R
\name{calcRSP}
\alias{calcRSP}
\title{Recreating RSP for a particular tracked animal}
\usage{
calcRSP(
  df.track,
  tz,
  distance,
  time.lapse,
  transition,
  er.ad,
  path.list,
  verbose
)
}
\arguments{
\item{df.track}{Detection data for that individual as imported using RSPete.}

\item{tz}{Time zone of the study area.}

\item{distance}{Maximum distance between RSP locations.}

\item{time.lapse}{Time lapse in minutes to be considered for consecutive detections at the same station.}

\item{transition}{TransitionLayer object as returned by LTDpath.}

\item{er.ad}{Incremental error per additional RSP point.}

\item{path.list}{A list of previously calculated paths.}

\item{verbose}{Logical: If TRUE, detailed messages and progression are displayed. Otherwise, a single progress bar is shown.}
}
\value{
A dataframe with the RSP estimations for all identified tracks for that animal.
}
\description{
Estimates the RSP individually for all tracks of a particular animal.
}
