% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-cfa.R
\name{CFA}
\alias{CFA}
\title{Testing measurement & structural models for dichotomous and polytomous data}
\usage{
CFA()
}
\value{
No return value, opens web browser and loads shiny application
}
\description{
Testing measurement & structural models for dichotomous and polytomous data
}
\examples{
\dontrun{CFA()}
}
