% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_check.R
\name{checkDupTimestamps}
\alias{checkDupTimestamps}
\title{Identify and remove duplicated timestamps}
\usage{
checkDupTimestamps(input, group, verbose = TRUE)
}
\arguments{
\item{input}{The detections to be checked}

\item{group}{The group being analysed, used solely for messaging purposes}
}
\value{
The detections without duplicated timestamps
}
\description{
Identify and remove duplicated timestamps
}
\keyword{internal}
