% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{checkGroupQuality}
\alias{checkGroupQuality}
\title{Performs a series of quality checks on the detection data.}
\usage{
checkGroupQuality(input, UTM.zone, verbose = TRUE)
}
\arguments{
\item{input}{The detections list}

\item{UTM.zone}{UTM zone of the study area.}

\item{verbose}{Logical: If TRUE, detailed check messages are displayed. Otherwise, only a summary is displayed.}
}
\value{
The detections which can be used for dbbmm
}
\description{
Performs a series of quality checks on the detection data.
}
\keyword{internal}
