% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{checkLocations}
\alias{checkLocations}
\title{Check location quality for the RSP output}
\usage{
checkLocations(input)
}
\arguments{
\item{input}{The output of runRSP.}
}
\value{
A plot showing the RSP locations by tracked group.
}
\description{
This function can be used to verify whether all RSP estimated positions were placed within the water along the 
study area.
}
