% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{checkSpatialWater}
\alias{checkSpatialWater}
\title{Check all receiver stations are in the water}
\usage{
checkSpatialWater(input, base.raster)
}
\arguments{
\item{input}{The spatial file from actel analysis.}

\item{base.raster}{Raster file from the study area defining land (1) and water (0) regions.}
}
\value{
A TRUE/FALSE value for validating in-water locations.
}
\description{
Verify that receivers are in-water within the base raster provided
}
