% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_check.R
\name{checkTrackPoints}
\alias{checkTrackPoints}
\title{Exclude tracks with less than 8 detections}
\usage{
checkTrackPoints(input, group, verbose = TRUE)
}
\arguments{
\item{input}{The detections to be checked}

\item{group}{The group being analysed, used solely for messaging purposes}
}
\value{
The detections for tracks with more than 8 detections
}
\description{
Exclude tracks with less than 8 detections
}
\keyword{internal}
