% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{checkTrackTimes}
\alias{checkTrackTimes}
\title{Exclude tracks shorter than 30 minutes:}
\usage{
checkTrackTimes(input, group, verbose = TRUE)
}
\arguments{
\item{input}{The detections to be checked}

\item{group}{The group being analysed, used solely for messaging purposes}
}
\value{
The detections for tracks longer than 30 minutes
}
\description{
Exclude tracks shorter than 30 minutes:
}
\keyword{internal}
