% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_helpers.R
\name{clipLand}
\alias{clipLand}
\title{Clip land areas from the dbbmm output}
\usage{
clipLand(input, base.raster)
}
\arguments{
\item{input}{The dbbmm}

\item{base.raster}{the base raster}
}
\value{
the dbbmm rasters with the land areas cut out
}
\description{
Clip land areas from the dbbmm output
}
\keyword{internal}
