% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_helpers.R
\name{compileTrackInfo}
\alias{compileTrackInfo}
\title{compiles summary information for the tracks used in the dbbmm}
\usage{
compileTrackInfo(group.list)
}
\arguments{
\item{group.list}{the lists of valid detections}
}
\value{
a list of valid track summaries
}
\description{
compiles summary information for the tracks used in the dbbmm
}
\keyword{internal}
