% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{densDetec}
\alias{densDetec}
\title{Density plot of elapsed times between consecutive acoustic detections}
\usage{
densDetec(input, type = "Total")
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}

\item{type}{Character vector (Total, Group) defining the level to which calculate density distributions.}
}
\value{
Density plots of hours elapsed between consecutive acoustic detections.
}
\description{
Generates a density plot for inspecting the distribution of elapsed times (in hours) between all consecutive
acustic detections. By default the plot is created including all monitored groups and transmitters. Alternatively,
can be set to be performed at group level using the type argument.
}
