% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{dist.calc}
\alias{dist.calc}
\title{Calculates the total distances travelled}
\usage{
dist.calc(input)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}
}
\value{
A dataframe of the total distances travelled using RSP and Receiver tracks.
}
\description{
Calculate the total distances travelled during each RSP track identified.
}
