% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-fa.R
\name{FA}
\alias{FA}
\title{Run exploratory factor analysis for dichotomous and polytomous data}
\usage{
FA()
}
\value{
No return value, opens web browser and loads shiny application
}
\description{
Run exploratory factor analysis for dichotomous and polytomous data
}
\examples{
\dontrun{FA()}
}
