% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getAreas}
\alias{getAreas}
\title{Calculate water areas per group or track}
\usage{
getAreas(input, type = c("group", "track"), breaks = c(0.5, 0.95))
}
\arguments{
\item{input}{The output of \code{\link{dynBBMM}}}

\item{type}{one of "group" or "track". If set to "track", UD rasters for each track are also supplied.}

\item{breaks}{The contours for calculating usage areas in squared metres. By default the 95\% and 50\% contours are used.}
}
\value{
A list of areas per track, per group
}
\description{
Calculate water areas per group or track
}
\examples{
\donttest{
# Import river shapefile
water <- actel::shapeToRaster(shape = paste0(system.file(package = "RSP"), "/River_latlon.shp"), 
size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Run dynamic Brownian Bridge Movement Model (dBBMM)
dbbmm.data <- dynBBMM(input = rsp.data, base.raster = water, UTM = 56)

# Get dBBMM areas at group level
areas.group <- getAreas(input = dbbmm.data, type = "group", breaks = c(0.5, 0.95))
}

}
