% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getCentroids}
\alias{getCentroids}
\title{Get centroid locations of dBBMM}
\usage{
getCentroids(input, areas, type, level, group, UTM)
}
\arguments{
\item{input}{The output of \code{\link{dynBBMM}}.}

\item{areas}{The output of \code{\link{getAreas}}.}

\item{type}{Character vector specifying the type of getAreas analysis performed: "group" or "track".}

\item{level}{Numeric vector defining the contour level of dBBMM of interest to extract the centroid positions.}

\item{group}{Character vector defining the group of interest for the analysis, when getAreas is of type "group".}

\item{UTM}{Numeric vector representing the UTM zone of the study area.}
}
\value{
A dataframe containing the centroid positions per each timeslot
}
\description{
When a timeslot dBBMM analysis is conducted, this function can be used to obtain 
centroid latitude and longitude locations between all utilization distribution
contours at group or track level.
}
\examples{
\donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Run dynamic Brownian Bridge Movement Model (dBBMM) with timeslots:
dbbmm.data <- dynBBMM(input = rsp.data, base.raster = water, UTM = 56, timeframe = 2)

# Get dBBMM areas at group level
areas.group <- getAreas(dbbmm.data, type = "group", breaks = c(0.5, 0.95))

# Obtaing centroid coordinate locations of dBBMM:
df.centroid <- getCentroids(input = dbbmm.data, areas = areas.group, type = "group",
   level = 0.95, group = "G1", UTM = 56)
}

}
