% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{getDistance}
\alias{getDistance}
\title{Get total distances travelled}
\usage{
getDistance(input)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}
}
\value{
A dataframe containing the total distances travelled during each RSP track.
}
\description{
Obtain the total distances travelled (in kilometers) for the tracked animals, using only the 
receiver locations and also adding the RSP positions.
}
