% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getDistances}
\alias{getDistances}
\title{Get total distances travelled}
\usage{
getDistances(input)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}
}
\value{
A dataframe containing the total distances travelled during each RSP track.
}
\description{
Obtain the total distances travelled (in kilometres) for the tracked animals, using only the 
receiver locations and also adding the RSP positions.
}
