% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getDistPoint}
\alias{getDistPoint}
\title{Calculate in-water distances from RSP locations to a point of reference}
\usage{
getDistPoint(input, point, t.layer, transmitter = NULL)
}
\arguments{
\item{input}{The output of \code{\link{runRSP}}}

\item{point}{Point of reference (lon, lat) to where distances will be calculated.}

\item{t.layer}{A transition layer. Can be calculated using the function \code{\link[actel]{transitionLayer}}.}

\item{transmitter}{The animal(s) of interest for calculating distances. If not specified, by default, distances will be calculated for all animals in the dataset.}
}
\value{
The RSP detections object with a distance (in metres) column appended.
}
\description{
Calculate in-water distances from RSP locations to a point of reference
}
\examples{
\donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Calculate distances to a point of reference
df.dist <- getDistPoint(input = rsp.data, point = c(151.0291, -33.81771), t.layer = tl)
}

}
