% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{getMeta}
\alias{getMeta}
\title{Get metadata from total dBBMM}
\usage{
getMeta(input, group)
}
\arguments{
\item{input}{The total dBBMM as returned by the dynBBMM function.}

\item{group}{Group of interest.}
}
\value{
A data frame with metadata for the group of interest.
}
\description{
Extracts movement metadata from the total dynamic Brownian Bridge Movement Model (dBBMM) for a particular group of interest.
}
