% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getOverlaps}
\alias{getOverlaps}
\title{Calculate overlaps between different groups}
\usage{
getOverlaps(input)
}
\arguments{
\item{input}{The output of \code{\link{getAreas}}}
}
\value{
A list of Overlaps (per timeslot if relevant), as well as the respective overlap rasters.
}
\description{
Calculate overlaps between different groups
}
\examples{
\donttest{
# Import river shapefile
water <- actel::shapeToRaster(shape = paste0(system.file(package = "RSP"), "/River_latlon.shp"), 
size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Run dynamic Brownian Bridge Movement Model (dBBMM)
dbbmm.data <- dynBBMM(input = rsp.data, base.raster = water, UTM = 56)

# Get dBBMM areas at group level
areas.group <- getAreas(dbbmm.data, type = "group", breaks = c(0.5, 0.95))

# Get overlaps between groups
overlap.data <- getOverlaps(areas.group)
}

}
