% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{getUTM}
\alias{getUTM}
\title{Convert WGS84 co-ords to UTM}
\usage{
getUTM(input, UTM.zone)
}
\arguments{
\item{input}{The detections to be checked}

\item{UTM.zone}{the UTM UTM.zone of the study area}
}
\value{
The detections with UTM coordinates
}
\description{
Convert WGS84 co-ords to UTM
}
\keyword{internal}
