% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{getWaterAreas}
\alias{getWaterAreas}
\title{Calculate water areas per track}
\usage{
getWaterAreas(dbbmm.rasters, base.raster, breaks)
}
\arguments{
\item{base.raster}{Path to the raster file from the study area.}

\item{breaks}{The contours for calculating usage areas in squared meters. By default the 95% and 50% contours are used.}

\item{dbbmm}{the results of the dBBMM calculation}
}
\value{
A list of areas per track, per group
}
\description{
Calculate water areas per track
}
\keyword{internal}
