% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{groupDetections}
\alias{groupDetections}
\title{Prepare detections for the dBBMM}
\usage{
groupDetections(detections, tz.study.area, bio, UTM.zone, timeframe = NULL)
}
\arguments{
\item{detections}{a list of detections per fish}

\item{tz.study.area}{the time UTM.zone of the study area}

\item{UTM.zone}{the UTM UTM.zone of the study area}
}
\value{
the detections grouped by group
}
\description{
Joins the detections by group.
}
\keyword{internal}
