% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input.example}
\alias{input.example}
\title{Example acoustic data filtered with actel}
\format{
A list of outputs from explore()
}
\source{
Data collected by the authors.
}
\usage{
input.example
}
\description{
A list with the results of an explore analysis ran for two bull sharks in Sydney Harbour.
}
\keyword{internal}
