% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-internal.R
\name{INTERNAL}
\alias{INTERNAL}
\title{Run exploratory factor analysis for dichotomous and polytomous data}
\usage{
INTERNAL()
}
\value{
No return value, opens web browser and loads shiny application
}
\description{
Run exploratory factor analysis for dichotomous and polytomous data
}
\examples{
\dontrun{FA()}
}
