% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-irt.R
\name{IRT}
\alias{IRT}
\title{Item calibration according to item response theory models}
\usage{
IRT()
}
\value{
No return value, opens web browser and loads shiny application
}
\description{
Item calibration according to item response theory models
}
\examples{
\dontrun{IRT()}
}
