% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-itemAnalysis.R
\name{ITEMAN}
\alias{ITEMAN}
\title{Item and test statistics based on classical test theory,}
\usage{
ITEMAN()
}
\value{
No return value, opens web browser and loads shiny application
}
\description{
Item and test statistics based on classical test theory,
}
\examples{
\dontrun{ITEMAN()}
}
