% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{loadRaster}
\alias{loadRaster}
\title{Load Raster for dBBMM}
\usage{
loadRaster(base.raster, UTM.zone)
}
\arguments{
\item{base.raster}{the name of the raster file}

\item{UTM.zone}{the UTM UTM.zone of the study area}
}
\value{
the raster object.
}
\description{
A simpler version of SPBDraster, as not as many conversions are needed.
}
\keyword{internal}
