% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{LonLatToUTM}
\alias{LonLatToUTM}
\title{Converts coordinates to UTM projection}
\usage{
LonLatToUTM(x, y, UTM.zone)
}
\arguments{
\item{x}{Vector of Longitudes in decimal degrees.}

\item{y}{Vector of Latitudes in decimal degrees.}

\item{UTM.zone}{UTM UTM.zone of input locations.}
}
\value{
Dataframe with the converted coordinates in UTM.
}
\description{
Convert Coordinate Reference System (CRS) from lonlat to 
the UTM projection and meter units, as required for calculating the dynamic Brownian Bridge 
Movement Model.
}
