% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pca.R
\name{PCA}
\alias{PCA}
\title{Run principal component analysis for dichotomous and polytomous data}
\usage{
PCA()
}
\value{
No return value, opens web browser and loads shiny application
}
\description{
Run principal component analysis for dichotomous and polytomous data
}
\examples{
\dontrun{PCA()}
}
