% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAreas}
\alias{plotAreas}
\title{Plot areas}
\usage{
plotAreas(
  areas,
  base.raster,
  group,
  timeslot,
  title = NULL,
  col,
  land.col = "#BABCBF80"
)
}
\arguments{
\item{areas}{The areas object used to calculate the space use areas at group level.}

\item{base.raster}{The raster used in the dbbmm calculations.}

\item{group}{Character vector indicating the group to be displayed.}

\item{timeslot}{The timeslot to be displayed. Only relevant for timeslot dbbmms.}

\item{title}{Plot title.}

\item{col}{Character vector of colours to be used in the plot (same length as the number of contour levels).}

\item{land.col}{Colour of the land masses. Defaults to semi-transparent grey.}
}
\value{
A plot of the overlapping areas between two groups.
}
\description{
Plot areas for a specific group and, if relevant, track and timeslot.
If the base raster is in a geographic coordinate system, plotAreas will attempt to convert the dbbmm results
to that same geographic system, so everything falls in place.
}
