% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDensities}
\alias{plotDensities}
\title{Density plot of elapsed times between consecutive acoustic detections}
\usage{
plotDensities(input, group)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}

\item{group}{Character vector defining the group to which calculate density distributions. By default, density is calculated for all animals and groups tracked.}
}
\value{
Density plots of hours elapsed between consecutive acoustic detections.
}
\description{
Generates a density plot for inspecting the distribution of elapsed times (in hours) between all consecutive
acustic detections. By default the plot is created including all monitored groups and transmitters. Alternatively,
can be set to be performed at group level using the type argument.
}
\examples{
\donttest{
# Import river shapefile
water <- actel::shapeToRaster(shape = paste0(system.file(package = "RSP"), "/River_latlon.shp"), 
size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Plot distribution of acoustic detections
plotDensities(rsp.data, group = "G1")
}

}
