% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotDetec}
\alias{plotDetec}
\title{Visualize RSP x Receiver total number of individual locations}
\usage{
plotDetec(input)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}
}
\value{
A barplot of total number of locations as a function of location type (Loc.type).
}
\description{
Compare the outputs of total number of individual location data for each tracked animal, 
using only receiver locations and adding the RSP positions.
}
