% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotDistance}
\alias{plotDistance}
\title{Plot total distances travelled}
\usage{
plotDistance(input, Group = FALSE)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}

\item{Group}{If TRUE, plots are returned individually for each tracked group.}
}
\value{
A barplot of total distances travelled as a function of location type (Loc.type) and the distances travelled during each RSP track.
}
\description{
Compare the outputs of total distances travelled (in kilometers) for the tracked animals, using only the 
receiver locations and adding the RSP positions. Data on the total distances travelled are stored in the 
'distances' objtect.
}
