% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotDistances}
\alias{plotDistances}
\title{Analyze RSP x Receiver total distances travelled}
\usage{
plotDistances(input)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}
}
\value{
A barplot of total distances travelled as a function of location type (Loc.type).
}
\description{
Compare the outputs of total distances travelled (in kilometers) for each tracked animal, 
using only receiver locations and adding the RSP positions.
}
