% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotOverlap}
\alias{plotOverlap}
\title{Plot orverlapping contours}
\usage{
plotOverlap(
  input,
  timeslot = NULL,
  stations = FALSE,
  level = 0.95,
  main = NULL,
  color.plot = NULL,
  land.col = "#BABCBF",
  store = FALSE
)
}
\arguments{
\item{input}{Dynamic Brownian Bridge Movement Model object as returned by dynBBMM.}

\item{timeslot}{The timeslot to be plotted.}

\item{stations}{Should receiver stations be added to the graph. Default is TRUE.}

\item{level}{Numeric vector of the use area to plot. By default the .95 areas will be returned.}

\item{main}{Character vector of the plot title. By default, the temporal window is returned in the title.}

\item{color.plot}{Character vector of contour colours in the following order: 1) group 1, 2) group 2, and 3) overlap.}

\item{land.col}{Color of the land mass.}

\item{store}{Logical: If TRUE, a list of plots is returned.}
}
\value{
dynamic Brownian Bridge Movement Model plot.
}
\description{
Plot specific dBBMM contours. By default, the contour is chosen to be 95\%.
}
