% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRaster}
\alias{plotRaster}
\title{Check input data quality for the RSP analysis}
\usage{
plotRaster(
  input,
  base.raster,
  coord.x,
  coord.y,
  size = 1,
  land.col = "#BABCBF80"
)
}
\arguments{
\item{input}{Either a data frame containing the coordinates of the stations or the output of one of 
\code{\link[actel]{actel}}'s main functions (\code{\link[actel]{explore}}, \code{\link[actel]{migration}} 
or \code{\link[actel]{residency}}).}

\item{base.raster}{Raster object. Imported for example using \code{\link[actel]{loadShape}}.}

\item{coord.x}{The names of the columns containing the x and y positions of the stations in the spatial object.}

\item{coord.y}{The names of the columns containing the x and y positions of the stations in the spatial object.}

\item{size}{The size of the station dots}

\item{land.col}{Colour of the land masses. Defaults to semi-transparent grey.}
}
\value{
A plot of your base raster extent and the receiver locations.
}
\description{
If you are reading this it's because RSP failed to detect all of your receivers within the base raster provided, 
or any of your receiver location was found to be in land. This function allows you to visually identify the station(s) 
with problem. Please either extend your raster to include all stations or fix receiver locations to be in-water.
}
