% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotRSP}
\alias{plotRSP}
\title{Plot comparison tracks}
\usage{
plotRSP(
  input,
  tag,
  display = c("Receiver", "RSP", "Both"),
  type = c("lines", "points")
)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}

\item{type}{Type of tracking plot to be generated: Receiver, RSP or Both.}

\item{animal}{Select a particular animal to plot.}

\item{base.raster}{Raster file of the study area.}
}
\value{
Tracking plot for the interest animal.
}
\description{
Compare animal tracks using receiver locations and RSP tracks.
}
