% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{prepareDetections}
\alias{prepareDetections}
\title{Import detection data as sorted format}
\usage{
prepareDetections(detections, spatial)
}
\arguments{
\item{detections}{A list of detections provived by an actel function.}

\item{spatial}{A list of spatial objects in the study area}
}
\value{
A standardized dataframe to be used for RSP calculation.
}
\description{
Open and sort the detections dataset for applying RSP estimation, using the tagging data to assign 
species names and indexes for each tracked animal. Also coverts the UTC date and time column to the 
local time zone of the study area.
}
