% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending.R
\name{rspDetec}
\alias{rspDetec}
\title{Extract RSP x Receiver total number of individual locations}
\usage{
rspDetec(input)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}
}
\value{
A dataframe of total number of locations as a function of location type (Loc.type).
}
\description{
Calculate total number of individual location data for each tracked animal, 
using only receiver locations and adding the RSP positions.
}
