% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{rspDist}
\alias{rspDist}
\title{Extract RSP x Receiver total distances travelled}
\usage{
rspDist(input)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}
}
\value{
A dataframe of total distances travelled as a function of location type (Loc.type).
}
\description{
Calculate the total distances travelled (in kilometers) for each tracked animal, 
using only receiver locations and adding the RSP positions.
}
