% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{RSPtransition}
\alias{RSPtransition}
\title{TransitionLayer object for LTD calculation}
\usage{
RSPtransition(raster.hab = "shapefile.grd")
}
\arguments{
\item{raster.hab}{RasterLayer object from the study area.}
}
\value{
TransitionLayer object for the study area.
}
\description{
Generates a TransitionLayer object from the study area for calculation of shortest paths.
}
