% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{saveTrackInfo}
\alias{saveTrackInfo}
\title{Compile a summary for each track (and slot, if timeframes were used)}
\usage{
saveTrackInfo(input, water, tz.study.area)
}
\arguments{
\item{input}{The detections list used for the dbbmm}

\item{water}{The water areas computed by getWaterAreas}

\item{tz.study.area}{TimeUTM.zone of the study area.}
}
\value{
a summary list with track information
}
\description{
Compile a summary for each track (and slot, if timeframes were used)
}
\keyword{internal}
