% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-dataSim.R
\name{SIMDATA}
\alias{SIMDATA}
\title{Generate simulated data according to IRT for dichotomous and polytomous data
Generate multidimensional data for factor analysis
# param options(java.parameters = "-Xmx8000m")}
\usage{
SIMDATA()
}
\value{
No return value, opens web browser and loads shiny application
}
\description{
Generate simulated data according to IRT for dichotomous and polytomous data
Generate multidimensional data for factor analysis
# param options(java.parameters = "-Xmx8000m")
}
\examples{
\dontrun{SIMDATA()}
}
