% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{timeInterval}
\alias{timeInterval}
\title{Calculate temporal differences between consecutive detections}
\usage{
timeInterval(input)
}
\arguments{
\item{input}{Detection dataset imported using the RSPete function.}
}
\value{
A dataframe with temporal differences in days between consecutive detection dates.
}
\description{
Calculates temporal differences in days between consecutive detection dates.
}
