% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{toUTM}
\alias{toUTM}
\title{Converts coordinates to UTM projection}
\usage{
toUTM(input, UTM, crs)
}
\arguments{
\item{input}{The detections data frame}

\item{UTM}{the UTM zone chosen by the user}

\item{crs}{The original coordinate system}
}
\value{
Dataframe with the converted coordinates in UTM.
}
\description{
Convert Coordinate Reference System (CRS) from lonlat to 
metric UTM projection, as required for calculating the dynamic Brownian Bridge 
Movement Model.
}
\keyword{internal}
