% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{trackNames}
\alias{trackNames}
\title{Identify potential fine-scale data for analysis}
\usage{
trackNames(detections, input, maximum.time = 1)
}
\arguments{
\item{detections}{# HF: missing variable}

\item{input}{Detection dates and temporal lags in days as returned by timeInterval.}

\item{maximum.time}{Temporal lag in days to be considered for the fine-scale tracking. Default is to consider 1-day intervals.}
}
\value{
A dataframe with identified and named individual tracks for RSP estimation.
}
\description{
Identifies fine-scale data among total detection dataset to be used for RSP estimation. Tracks are 
then named based on the interval between consecutive detection dates.
}
