% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_helpers.R
\name{trimDetections}
\alias{trimDetections}
\title{Select specific transmitters to analyze}
\usage{
trimDetections(detections, tags = NULL)
}
\arguments{
\item{detections}{The detections data frame, provided by one of the main actel functions (explore, migrate, residency of spbd)}

\item{tags}{A list of transmitters to be analysed.}
}
\value{
the trimmed detections list
}
\description{
If the user specifies transmitters, return only the detections for those tags.
}
\keyword{internal}
