% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{addStations}
\alias{addStations}
\title{Add receiver stations to an existing plot}
\usage{
addStations(input, shape = 21, size = 1.5, colour = "white", fill = "black")
}
\arguments{
\item{input}{The output of \code{\link{runRSP}} or \code{\link{dynBBMM}}}

\item{shape}{The shape of the points}

\item{size}{The size of the points}

\item{colour}{The colour of the points}

\item{fill}{The fill of the points}
}
\value{
A ggplot with stations
}
\description{
Add receiver stations to an existing plot
}
