% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{calculateDBBMM}
\alias{calculateDBBMM}
\title{Calculate the dBBMM for each group}
\usage{
calculateDBBMM(input, crs, base.raster)
}
\arguments{
\item{input}{The detections to be used as input for the model}

\item{base.raster}{The raster object}
}
\value{
A list of dBBMM's per group
}
\description{
Calculate the dBBMM for each group
}
\keyword{internal}
