% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{checkSpatial}
\alias{checkSpatial}
\title{Check all receiver stations are in the water}
\usage{
checkSpatial(spatial, base.raster)
}
\arguments{
\item{spatial}{The spatial file from actel analysis.}

\item{base.raster}{Raster file from the study area defining land (1) and water (0) regions.}
}
\value{
A standardized dataframe to be used for RSP calculation.
}
\description{
Verify that receivers are in-water within the base raster provided
}
