% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIMOS.R
\name{convertIMOS}
\alias{convertIMOS}
\title{Convert IMOS files to be analysed with RSP}
\usage{
convertIMOS(det, rmeta, tmeta, meas)
}
\arguments{
\item{det}{Path to IMOS detections file, or QC object processed with the remora package. Please see the package vignettes for details.}

\item{rmeta}{Path to IMOS receiver deployment metadata file.}

\item{tmeta}{Path to IMOS tansmitter deployment metadata file.}

\item{meas}{Path to IMOS animal measurements file.}
}
\value{
Returns an RSP object that can be used as input for the runRSP function.
}
\description{
This function converts between the IMOS data formats, and quality controlled (QC) objects, to the 
appropriate format required to be analysed with RSP.
}
