% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{dynBBMM}
\alias{dynBBMM}
\title{Total dynamic Brownian Bridge Movement Model}
\usage{
dynBBMM(
  input,
  UTM.zone,
  tags = NULL,
  breaks = c(0.95, 0.5),
  timeframe = NULL,
  debug = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input}{The output of runRSP.}

\item{UTM.zone}{UTM zone of the study area.}

\item{tags}{Vector of transmitters to be analyzed. By default all transmitters from runRSP will be analised.}

\item{breaks}{The contours for calculating usage areas in squared meters. By default the 95\% and 50\% contours are used.}

\item{timeframe}{Temporal window size for fine-scale dBBMM in hours. If left NULL, a single dBBMM is calculated for the whole period.}

\item{debug}{Logical: If TRUE, the function progress is saved to an RData file.}

\item{verbose}{Logical: If TRUE, detailed check messages are displayed. Otherwise, only a summary is displayed.}
}
\value{
List of calculated dBBMMs and metadata on each track used for the modelling.
}
\description{
Calculates dynamic Brownian Bridge Movement Model (dBBMM) for each track and transmitter. Tracks shorter than 30 minutes
are automatically identified and not included in the analysis.
}
