% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{dynBBMM}
\alias{dynBBMM}
\title{Total dynamic Brownian Bridge Movement Model}
\usage{
dynBBMM(
  input,
  base.raster,
  tags = NULL,
  start.time,
  stop.time,
  timeframe = NULL,
  UTM,
  debug = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input}{The output of runRSP.}

\item{base.raster}{The water raster of the study area. For example the output of \code{\link[actel]{loadShape}}.}

\item{tags}{Vector of transmitters to be analysed. By default all transmitters from runRSP will be analysed.}

\item{start.time}{Sets the start point for analysis (format = "Y-m-d H:M:S").}

\item{stop.time}{Sets the stop point for analysis (format = "Y-m-d H:M:S").}

\item{timeframe}{Temporal window size for fine-scale dBBMM in hours. If left NULL, a single dBBMM is calculated for the whole period.}

\item{UTM}{The UTM zone of the study area. Only relevant if a latlon-to-metric conversion is required.}

\item{debug}{Logical: If TRUE, the function progress is saved to an RData file.}

\item{verbose}{Logical: If TRUE, detailed check messages are displayed. Otherwise, only a summary is displayed.}
}
\value{
List of calculated dBBMMs and metadata on each track used for the modelling.
}
\description{
Calculates dynamic Brownian Bridge Movement Model (dBBMM) for each track and transmitter. Tracks shorter than 30 minutes
are automatically identified and not included in the analysis.
}
\examples{
\donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Run dynamic Brownian Bridge Movement Model (dBBMM)
dbbmm.data <- dynBBMM(input = rsp.data, base.raster = water, UTM = 56)
}

}
