% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getDistances}
\alias{getDistances}
\title{Get total distances travelled}
\usage{
getDistances(input, t.layer)
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}

\item{t.layer}{A transition layer. Can be calculated using the function \code{\link[actel]{transitionLayer}}.}
}
\value{
A dataframe containing the total distances travelled during each RSP track.
}
\description{
Obtain the total distances travelled (in metres) for the tracked animals, using only the 
receiver locations and also adding the RSP positions.
}
\examples{
\donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Calculate distances travelled
distance.data <- getDistances(rsp.data, t.layer = tl)
}

}
