% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getOverlaps}
\alias{getOverlaps}
\title{Calculate overlaps between different groups}
\usage{
getOverlaps(input)
}
\arguments{
\item{input}{The output of \code{\link{getAreas}}}
}
\value{
A list of Overlaps (per timeslot if relevant), as well as the respective overlap rasters.
}
\description{
Calculate overlaps between different groups
}
