% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynBBMM.R
\name{getOverlaps}
\alias{getOverlaps}
\title{Calculate overlaps between different groups}
\usage{
getOverlaps(input, breaks)
}
\arguments{
\item{input}{The output of \code{\link{dynBBMM}}}

\item{breaks}{The contours for calculating usage areas in squared metres. By default the 95\% and 50\% contours are used.}
}
\value{
A list of areas per track, per group
}
\description{
Calculate overlaps between different groups
}
\keyword{export}
