% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{includeRSP}
\alias{includeRSP}
\title{Recreating RSP for all tracked animals}
\usage{
includeRSP(
  detections,
  transition,
  tz.study.area,
  distance,
  time.lapse,
  er.ad,
  debug = FALSE
)
}
\arguments{
\item{detections}{Detection data for that individual as imported using RSPete.}

\item{transition}{TransitionLayer object as returned by LTDpath.}

\item{tz.study.area}{Timezone of the study area.}

\item{distance}{Maximum distance between RSP locations.}

\item{time.lapse}{Time lapse in minutes to be considered for consecutive detections at the same station.}

\item{er.ad}{Incremental error per additional RSP point.}
}
\value{
A list with the RSP estimations of individual tracks per transmitter.
}
\description{
Automatically estimates the RSP for all tracked individuals within a particular study area.
}
