% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{nameTracks}
\alias{nameTracks}
\title{Identify potential fine-scale data for analysis}
\usage{
nameTracks(detections, maximum.time = 24)
}
\arguments{
\item{detections}{Detections data frame}

\item{maximum.time}{Temporal lag in hours to be considered for the fine-scale tracking. Default is to consider 1-day intervals.}
}
\value{
A dataframe with identified and named individual tracks for RSP estimation.
}
\description{
Identifies fine-scale data among total detection dataset to be used for RSP estimation. Tracks are 
then named based on the interval between consecutive detection dates.
}
