% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotContours}
\alias{plotContours}
\title{Plot dynamic Brownian Bridge Movement Models (dBBMM)}
\usage{
plotContours(
  input,
  group,
  track = NULL,
  timeslot = NULL,
  stations = FALSE,
  levels = c(0.99, 0.95, 0.75, 0.5, 0.25),
  main = NULL,
  land.col = "#BABCBF"
)
}
\arguments{
\item{input}{Dynamic Brownian Bridge Movement Model object as returned by dynBBMM.}

\item{group}{Group/species of transmitters.}

\item{levels}{Numeric vector os use areas to plot. By default the 99%, 95%, 75%, 50% and 25% areas will be returned.}

\item{land.col}{Color of the land mass.}

\item{Track}{Transmitter and track names to plot.}

\item{Station}{Should receiver stations be added to the graph. Default is TRUE.}
}
\value{
dynamic Brownian Bridge Movement Model plot.
}
\description{
Plot specific dBBMM contours. By default, the inside contour (level1) is chosen to be the 50% 
and the outer (level2) to be the 95%.
}
