% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDistances}
\alias{plotDistances}
\title{Plot total distances travelled}
\usage{
plotDistances(input, group, compare = TRUE)
}
\arguments{
\item{input}{output of \code{\link{getDistances}}.}

\item{group}{Define a specific group to be plotted, rather than the overall results.}

\item{compare}{By default, a comparative plot is returned showing distances travelled with Receiver and RSP location
types. If FALSE, only the RSP total distances travelled will be returned.}
}
\value{
A barplot of total distances travelled as a function of location type (Loc.type) and the distances travelled during each RSP track.
}
\description{
Compare the outputs of total distances travelled (in kilometres) for the tracked animals, using only the 
receiver locations and adding the RSP positions. Data on the total distances travelled are stored in the 
'distances' objtect.
}
\examples{
\donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Calculate distances travelled
distance.data <- getDistances(rsp.data, t.layer = tl)

# Plot distances travelled
plotDistances(distance.data, group = "G1")
}

}
