% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotOverlaps}
\alias{plotOverlaps}
\title{Plot overlapping contours}
\usage{
plotOverlaps(
  overlaps,
  areas,
  base.raster,
  groups,
  timeslot,
  level,
  title = NULL,
  col,
  land.col = "#BABCBF80"
)
}
\arguments{
\item{overlaps}{An overlap object as returned by \code{\link{getOverlaps}}.}

\item{areas}{The areas object used to calculate the overlaps.}

\item{base.raster}{The raster used in the dbbmm calculations.}

\item{groups}{Character vector indicating the two groups to be displayed.}

\item{timeslot}{The timeslot to be displayed. Only relevant for timeslot dbbmms.}

\item{level}{Value of the use area to plot. Must match one the levels calculated in the overlaps.}

\item{title}{Plot title. By default, the names of the groups being compared are displayed.}

\item{col}{Character vector of three colours to be used in the plot (one for each group and one for the overlap).}

\item{land.col}{Colour of the land masses. Defaults to semi-transparent grey.}
}
\value{
A plot of the overlapping areas between two groups.
}
\description{
Plot specific dBBMM overlapping areas for a specific combination of groups and, if relevant, a specific timeslot.
If the base raster is in a geographic coordinate system, plotOverlaps will attempt to convert the dbbmm results
to that same geographic system, so everything falls in place.
}
