% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotOverlaps}
\alias{plotOverlaps}
\title{Plot orverlapping contours}
\usage{
plotOverlaps(
  input,
  base.raster,
  timeslot = NULL,
  stations = FALSE,
  level = 0.95,
  main = NULL,
  color.plot,
  land.col = "#BABCBF",
  store = FALSE
)
}
\arguments{
\item{input}{An overlap object as returned by getOverlaps.}

\item{base.raster}{Raster file from the study area defining land (1) and water (0) regions.}

\item{timeslot}{The timeslot to be plotted.}

\item{stations}{Should receiver stations be added to the graph. Default is TRUE.}

\item{level}{Numeric vector of the use area to plot. By default the .95 areas will be returned.}

\item{main}{Character vector of the plot title. By default, the temporal window is returned in the title.}

\item{color.plot}{Character vector of colours for overlapping area.}

\item{land.col}{Color of the land mass.}

\item{store}{Logical: If TRUE, a list of plots is returned.}
}
\value{
A plot of the overlapping areas between groups.
}
\description{
Plot specific dBBMM overlaping areas when multiple groups are tracked. The contour is automatically chosen to be 95\% if the 
default version of getOverlaps() was used. If overlaps were calculated at a different and specific contour, please provide it through the level argument.
}
