% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotRaster}
\alias{plotRaster}
\title{Check input data quality for the RSP analysis}
\usage{
plotRaster(input, base.raster, coord.x, coord.y, size)
}
\arguments{
\item{input}{The output of one of actel's main functions (explore, migration or residency)}

\item{base.raster}{Raster file from the study area defining land (1) and water (0) regions.}

\item{coord.x}{The names of the columns containing the x and y positions of the stations 
in the spatial object.}

\item{coord.y}{The names of the columns containing the x and y positions of the stations 
in the spatial object.}

\item{size}{The size of the station dots}
}
\value{
A plot of your base raster extent and the receiver locations.
}
\description{
If you are reading this it's because RSP failed to detect all of your receivers within the base raster provided, 
or any of your receiver location was found to be in land. This function allows you to visually identify the station(s) 
with problem. Please either extend your raster to include all stations or fix receiver locations to be in-water.
}
