% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotRSP}
\alias{plotRSP}
\title{Plot comparison tracks}
\usage{
plotRSP(input, tag, display = c("Receiver", "RSP", "Both"),
  type = c("lines", "points"))
}
\arguments{
\item{input}{RSP dataset as returned by RSP.}

\item{tag}{Select a particular animal to plot.}

\item{display}{Tracks to be displayed. One of Receiver, RSP or Both.}

\item{type}{Type of plot. One of lines or points}
}
\value{
Tracking plot for the interest animal.
}
\description{
Compare animal tracks using receiver locations and RSP tracks.
}
