% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTracks}
\alias{plotTracks}
\title{Plot the RSP tracks}
\usage{
plotTracks(
  input,
  base.raster,
  type = c("both", "points", "lines"),
  group,
  tag,
  track,
  size = c(0.33, 0.3),
  alpha = c(0.5, 0.5),
  land.col = "#BABCBF80"
)
}
\arguments{
\item{input}{The output of runRSP.}

\item{base.raster}{The raster used to generate the transition layer used in runRSP}

\item{type}{One of "points", "line" or "both". Defaults to "both", i.e. both lines and points are plotted for the
generated tracks.}

\item{group}{Choose a single group of fish to plot}

\item{tag}{Choose a single tag to plot}

\item{track}{If a single tag was chosen, you can use 'track' to define a specific track to be plotted.}

\item{size}{The size/width of the points and lines to be plotted. if type = "both", the line size will be the
one specified and the point size will be 10\% larger than the specified.}

\item{alpha}{One or two transparency values (for points and lines, respectively). For no transparency, alpha = 1.}

\item{land.col}{Colour of the land masses. Defaults to semi-transparent grey.}
}
\value{
A plot showing the RSP track locations.
}
\description{
This function can be used to plot a map of a particular RSP track of interest.
}
\examples{
\donttest{
# Import river shapefile
water <- actel::shapeToRaster(shape = paste0(system.file(package = "RSP"), "/River_latlon.shp"), 
size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")

# Plot a specific RSP track
plotTracks(rsp.data, base.raster = water, tag = "A69-9001-1111", track = "Track_1")
}

}
