% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plotTracks}
\alias{plotTracks}
\title{Check location quality for the RSP output}
\usage{
plotTracks(
  input,
  base.raster,
  type = c("both", "points", "lines"),
  group,
  tag,
  track,
  size = 0.3
)
}
\arguments{
\item{input}{The output of runRSP.}

\item{base.raster}{The raster used to generate the transition layer used in runRSP}

\item{type}{One of "points", "line" or "both". Defaults to "both", i.e. both lines and points are plotted for the
generated tracks.}

\item{group}{Choose a single group of fish to plot}

\item{tag}{Choose a single tag to plot}

\item{track}{If a single tag was chosen, you can use 'track' to define a specific track to be plotted.}

\item{size}{The size/width of the points and lines to be plotted. if type = "both", the line size will be the
one specified and the point size will be 10\% larger than the specified.}
}
\value{
A plot showing the RSP locations by tracked group.
}
\description{
This function can be used to verify whether all RSP estimated positions were placed within the water along the 
study area.
}
