% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{runRSP}
\alias{runRSP}
\title{Refined Shortest Path (RSP) between detections}
\usage{
runRSP(
  input,
  t.layer,
  coord.x,
  coord.y,
  distance = 250,
  time.lapse = 10,
  er.ad = NULL,
  maximum.time = 24,
  debug = FALSE
)
}
\arguments{
\item{input}{The output of one of \code{\link[actel]{actel}}'s main functions (\code{\link[actel]{explore}}, \code{\link[actel]{migration}} or \code{\link[actel]{residency}})}

\item{t.layer}{A transition layer calculated using the function \code{\link[actel]{transitionLayer}}}

\item{coord.x, coord.y}{The names of the columns containing the x and y positions of the stations 
in the spatial object.}

\item{distance}{Fixed distances in meters to add intermediate track locations. By default intermediate positions are added every 250 metres.}

\item{time.lapse}{Time lapse in minutes to be considered for consecutive detections at the same station.}

\item{er.ad}{By default, 5\% of the distance argument is used as the increment rate of the position errors for the estimated locations. Alternatively, can be defined by the user in meters.}

\item{maximum.time}{Temporal lag in hours to be considered for the fine-scale tracking. Default is to consider 1-day intervals.}

\item{debug}{Logical: If TRUE, the function progress is saved to an RData file.}
}
\value{
Returns a list of RSP tracks (as data frame) for each transmitter detected.
}
\description{
Estimates the RSP for a series of animals tracked with acoustic transmitters. Intermediate
locations between consecutive acoustic detections (either on the same or different receivers) are estimated 
according to fixed distance (see distance for details) and temporal (see time.lapse for details) intervals. The error of estimated locations increase proportionally as 
the animal moves away from the first detection, and decreases as it approaches the second detection (see er.ad for details). If 
the animal is not detected for a long time (default is a daily absence), the detections are broken into a 
new track (see maximum.time for details).
}
