% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRSP.R
\name{runRSP}
\alias{runRSP}
\title{Calculate refined shortest paths between detections}
\usage{
runRSP(
  input,
  t.layer,
  coord.x,
  coord.y,
  distance = 250,
  tags = NULL,
  recaptures = FALSE,
  time.step = 10,
  min.time = 10,
  max.time = 24,
  er.ad,
  verbose = FALSE,
  debug = FALSE
)
}
\arguments{
\item{input}{The output of one of \code{\link[actel]{actel}}'s main functions (\code{\link[actel]{explore}}, \code{\link[actel]{migration}} or \code{\link[actel]{residency}})}

\item{t.layer}{A transition layer. Can be calculated using the function \code{\link[actel]{transitionLayer}}.}

\item{coord.x, coord.y}{The names of the columns containing the x and y positions of the stations in the spatial object.}

\item{distance}{Distance (in metres) by which RSP point should be spaced (between detections at different stations). Defaults to 250 metres.}

\item{tags}{Vector of transmitters for which to calculate RSP. By default all transmitters will be analysed.}

\item{recaptures}{Logical: if TRUE, a recapture.csv dataset containing the recapture locations of tracked animals will be included in the analysis.}

\item{time.step}{Time lapse (in minutes) between RSP points added between detections at the same station. Defaults to 10 minutes. Must not be larger than \code{min.time}.}

\item{min.time}{Minimum time required between receiver detections (in minutes) for RSP to be calculated. Default to 10 minutes.}

\item{max.time}{Maximum time allowed between receiver detections (in hours) for RSP to be calculated. Defaults to 24 hours.}

\item{er.ad}{Increment rate of the position errors for the estimated locations (in metres). If left unset, defaults to 5\% of the \code{distance} argument.}

\item{verbose}{Logical: If TRUE, detailed messages and progression are displayed. Otherwise, a single progress bar is shown.}

\item{debug}{Logical: If TRUE, the function progress is saved to an RData file.}
}
\value{
Returns a list of RSP tracks for each transmitter detected, as well as auxiliary information.
}
\description{
Estimates the RSP for a series of animals tracked with acoustic transmitters. Intermediate
locations between consecutive acoustic detections (either on the same or different receivers) are estimated 
according to the defined \code{distance} and \code{time.step} arguments. The error of estimated locations increase proportionally as 
the animal moves away from the first detection, and decreases as it approaches the second detection (argument \code{er.ad}). If 
the animal is not detected for a long time (default is a daily absence), the detections are broken into a 
new track (argument \code{max.time}).
}
\examples{
\donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")
}

}
