% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP.R
\name{runRSP}
\alias{runRSP}
\title{Refined Shortest Path (RSP) between detections}
\usage{
runRSP(
  input,
  base.raster,
  distance = 250,
  time.lapse = 10,
  er.ad = NULL,
  maximum.time = 24,
  debug = FALSE
)
}
\arguments{
\item{input}{The output of one of actel's main functions (explore, migration or residency)}

\item{base.raster}{Raster file from the study area defining land (1) and water (0) regions.}

\item{distance}{Fixed distances in meters to add intermediate track locations. By default intermediate positions are added every 250 m.}

\item{time.lapse}{Temporal limit for the RSP in minutes. Consecutive detections shorter than the time.lapse will not have interpolated positions.}

\item{er.ad}{By default, 5\% of the distance argument is used as the increment rate of the position erros for the estimated locations. Alternatively, can be defined by the user in meters.}

\item{maximum.time}{Temporal lag in hours to be considered for the fine-scale tracking. Default is to consider 1-day intervals.}

\item{debug}{Logical: If TRUE, the function progress is saved to an RData file.}
}
\value{
Returns a list of RSP tracks (as dataframe) for each transmitter detected.
}
\description{
Estimates the RSP for a series of animals tracked with acoustic transmitters. Intermediate
locations between consecutive acoustic detections (either on the same or different receivers) are estimated 
according to fixed distance (see distance for details) and temporal (see time.lapse for details) intervals. The error of estimated locations increase proportionally as 
the animal moves away from the first detection, and decreases as it approaches the second detection (see er.ad for details). If 
the animal is not detected for a long time (default is a daily absence), the detections are broken into a 
new track (see maximum.time for details).
}
