% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{suggestSize}
\alias{suggestSize}
\title{Suggest plot dimensions for a given raster}
\usage{
suggestSize(input, max)
}
\arguments{
\item{input}{The raster being plotted}

\item{max}{the desired size for the longest edge}
}
\value{
A width/height vector (rounded)
}
\description{
Suggest plot dimensions for a given raster
}
\examples{
\donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 

# Find suggested size to save projected map 
suggestSize(water, max = 10)
}

}
